working on getting tests passing
