+++ mmeds/tests/unit/test_spawn.py
@@ -4,27 +4,12 @@ from time import sleep
 
 from mmeds.authentication import add_user, remove_user
 from mmeds.qiime2 import Qiime2
-from mmeds.database import MetaDataUploader
+from mmeds.database import upload_metadata
 from mmeds.util import load_config
 
 import mmeds.config as fig
 import mmeds.secrets as sec
 import mmeds.spawn as sp
-
-
-def upload_metadata(args):
-    metadata, path, owner, access_code = args
-    with MetaDataUploader(metadata=metadata,
-                          path=path,
-                          study_type='qiime',
-                          study_name='Test_Spawn',
-                          reads_type='single_end',
-                          owner=fig.TEST_USER,
-                          temporary=False,
-                          testing=True) as up:
-        access_code,  email = up.import_metadata(for_reads=fig.TEST_READS,
-                                                 barcodes=fig.TEST_BARCODES,
-                                                 access_code=access_code)
 
 
 class SpawnTests(TestCase):
@@ -55,8 +40,8 @@ class SpawnTests(TestCase):
     @classmethod
     def tearDownClass(self):
         remove_user(fig.TEST_USER, testing=True)
-        #for new_dir in self.dirs:
-        #    rmtree(new_dir)
+        for new_dir in self.dirs:
+            rmtree(new_dir)
 
     def test_b_restart_analysis(self):
         """ Test restarting an analysis from a analysis doc. """
