+++ mmeds/util.py
@@ -1052,3 +1052,22 @@ def write_processes(process_codes):
                 # Add it's access code to the process log
                 all_codes.append('{}\t{}'.format(ptype, pdoc.analysis_code))
     fig.PROCESS_LOG.write_text('\n'.join(all_codes))
+
+
+
+
+def upload_metadata(args):
+    metadata, path, owner, reads_type, for_reads, rev_reads, barcodes, access_code = args
+    with MetaDataUploader(metadata=metadata,
+                          path=path,
+                          study_type='qiime',
+                          study_name='Test_Qiime',
+                          reads_type=reads_type,
+                          owner=fig.TEST_USER,
+                          temporary=False,
+                          public=False,
+                          testing=True) as up:
+        access_code, email = up.import_metadata(for_reads=for_reads,
+                                                rev_reads=rev_reads,
+                                                barcodes=barcodes,
+                                                access_code=access_code)
